src/newsletter_gen/tools/tools.py
from crewai_tools import BaseTool
import json
import re
from jinja2 import Template

class JobAnalysisTool(BaseTool):
    name: str = "Job Analysis Tool"
    description: str = "Extract requirements from job postings"
    
    def _run(self, job_text: str) -> dict:
        # This is a stub - the actual analysis will be done by the LLM
        # We're just providing a placeholder for the CrewAI framework
        return {"status": "Analysis requested", "job_text_length": len(job_text)}

class SkillsMatcherTool(BaseTool):
    name: str = "Skills Matcher Tool"
    description: str = "Match job requirements with candidate skills"
    
    def _run(self, job_requirements: dict) -> dict:
        # This is a stub - the actual matching will be done by the LLM
        # We're just providing a placeholder for the CrewAI framework
        return {"status": "Matching requested", "requirements_count": len(job_requirements)}

class HTMLGeneratorTool(BaseTool):
    name: str = "HTML Generator Tool"
    description: str = "Generate CV using template"
    
    def _run(self, matched_data: dict) -> str:
        # This is a stub - the actual HTML generation will be done by the LLM
        # We're just providing a placeholder for the CrewAI framework
        return f"HTML generation requested with {len(matched_data)} data points"