requirements.txt
# Core dependencies
crewai>=0.28.0
langchain>=0.1.0
langchain-openai>=0.0.2
playwright>=1.40.0
jinja2>=3.1.2
python-dotenv>=1.0.0

# UI
streamlit>=1.29.0

# Optional LLM providers (uncomment as needed)
# langchain-anthropic>=0.1.1
# langchain-groq>=0.1.0
# langchain-google-genai>=0.0.5

# For HTML/PDF conversion
# xhtml2pdf>=0.2.11
